ref(metrics): move metrics api tags endpoint to separate file
