+++ src/sentry/api/endpoints/organization_metrics_tags.py
@@ -1,0 +1,51 @@
+from rest_framework.exceptions import ParseError
+from rest_framework.request import Request
+from rest_framework.response import Response
+
+from sentry.api.api_owners import ApiOwner
+from sentry.api.api_publish_status import ApiPublishStatus
+from sentry.api.base import region_silo_endpoint
+from sentry.api.bases import OrganizationAndStaffPermission, OrganizationEndpoint
+from sentry.api.endpoints.organization_metrics import get_use_case_id
+from sentry.api.utils import get_date_range_from_params
+from sentry.exceptions import InvalidParams
+from sentry.snuba.metrics import DerivedMetricParseException, get_all_tags
+
+
+@region_silo_endpoint
+class OrganizationMetricsTagsEndpoint(OrganizationEndpoint):
+    publish_status = {
+        "GET": ApiPublishStatus.EXPERIMENTAL,
+    }
+    owner = ApiOwner.TELEMETRY_EXPERIENCE
+    permission_classes = (OrganizationAndStaffPermission,)
+
+    """Get list of tag names for this project
+
+    If the ``metric`` query param is provided, only tags for a certain metric
+    are provided.
+
+    If the ``metric`` query param is provided more than once, the *intersection*
+    of available tags is used.
+    """
+
+    def get(self, request: Request, organization) -> Response:
+        metric_names = request.GET.getlist("metric") or []
+        projects = self.get_projects(request, organization)
+        if not projects:
+            raise InvalidParams("You must supply at least one project to see the tag names")
+
+        start, end = get_date_range_from_params(request.GET)
+
+        try:
+            tags = get_all_tags(
+                projects=projects,
+                metric_names=metric_names,
+                use_case_id=get_use_case_id(request),
+                start=start,
+                end=end,
+            )
+        except (InvalidParams, DerivedMetricParseException) as exc:
+            raise (ParseError(detail=str(exc)))
+
+        return Response(tags, status=200)
